requirements.txt
crispy-bootstrap5==2024.2
django-allauth==0.61.1
django-ckeditor==6.6.1
whitenoise==6.6.0
pillow==10.2.0
gunicorn==21.2.0
dj-database-url==2.1.0
boto3==1.34.55
django-storages==1.14.2
djangorestframework==3.14.0
django-environ==0.11.2
markdown==3.5.2
django-cloudinary-storage==0.3.0
psycopg2-binary
django-csp==3.8
black==24.2.0.
faker==23.3.0
django-debug-toolbar==4.3.0
coverage==7.4.3
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
python-decouple
gunicorn